+++ CHANGELOG.md
@@ -5,6 +5,7 @@ the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## Unreleased
 
+- Fixed: Multiple checkbox/select options in Livewire forms were not properly handled.
 - Added: `chief:default-menus` command to add default menus to the database.
 - Added: Highlight active nav item in the admin panel.
 - Changed: ⚠ BC break `chief.menus` config is now structured differently. Each key is the type and value is its label.